Rename `TryFromJsonSyntax` into `TryFromJson`. Add `KindSet::as_conjunction`. Add `KindSet::as_disjunction`. Add `kind` module.
