+++ src/lib.rs
@@ -48,6 +48,8 @@ pub use code_map::CodeMap;
 pub use parse::Parse;
 pub mod print;
 pub use print::Print;
+pub mod kind;
+pub use kind::{Kind, KindSet};
 mod convert;
 mod macros;
 mod try_from;
@@ -86,236 +88,6 @@ pub const NUMBER_CAPACITY: usize = SMALL_STRING_CAPACITY;
 
 /// Number buffer.
 pub type NumberBuf = json_number::SmallNumberBuf<NUMBER_CAPACITY>;
-
-/// Value kind.
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug)]
-pub enum Kind {
-	Null,
-	Boolean,
-	Number,
-	String,
-	Array,
-	Object,
-}
-
-impl std::ops::BitOr for Kind {
-	type Output = KindSet;
-
-	fn bitor(self, other: Self) -> KindSet {
-		KindSet::from(self) | KindSet::from(other)
-	}
-}
-
-impl std::ops::BitOr<KindSet> for Kind {
-	type Output = KindSet;
-
-	fn bitor(self, other: KindSet) -> KindSet {
-		KindSet::from(self) | other
-	}
-}
-
-impl std::ops::BitAnd for Kind {
-	type Output = KindSet;
-
-	fn bitand(self, other: Self) -> KindSet {
-		KindSet::from(self) & KindSet::from(other)
-	}
-}
-
-impl std::ops::BitAnd<KindSet> for Kind {
-	type Output = KindSet;
-
-	fn bitand(self, other: KindSet) -> KindSet {
-		KindSet::from(self) & other
-	}
-}
-
-impl fmt::Display for Kind {
-	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		match self {
-			Self::Null => write!(f, "null"),
-			Self::Boolean => write!(f, "boolean"),
-			Self::Number => write!(f, "number"),
-			Self::String => write!(f, "string"),
-			Self::Array => write!(f, "array"),
-			Self::Object => write!(f, "object"),
-		}
-	}
-}
-
-macro_rules! kind_set {
-	($($id:ident ($const:ident): $mask:literal),*) => {
-		#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-		pub struct KindSet(u8);
-
-		impl KindSet {
-			$(
-				pub const $const: Self = Self($mask);
-			)*
-
-			pub fn len(&self) -> usize {
-				self.0.count_ones() as usize
-			}
-
-			pub fn is_empty(&self) -> bool {
-				self.0 == 0
-			}
-
-			pub fn iter(&self) -> KindSetIter {
-				KindSetIter(self.0)
-			}
-		}
-
-		impl std::ops::BitOr<Kind> for KindSet {
-			type Output = Self;
-
-			fn bitor(self, other: Kind) -> Self {
-				match other {
-					$(
-						Kind::$id => Self(self.0 | $mask)
-					),*
-				}
-			}
-		}
-
-		impl std::ops::BitOrAssign<Kind> for KindSet {
-			fn bitor_assign(&mut self, other: Kind) {
-				match other {
-					$(
-						Kind::$id => self.0 |= $mask
-					),*
-				}
-			}
-		}
-
-		impl std::ops::BitOr for KindSet {
-			type Output = Self;
-
-			fn bitor(self, other: Self) -> Self {
-				Self(self.0 | other.0)
-			}
-		}
-
-		impl std::ops::BitOrAssign for KindSet {
-			fn bitor_assign(&mut self, other: Self) {
-				self.0 |= other.0
-			}
-		}
-
-		impl std::ops::BitAnd<Kind> for KindSet {
-			type Output = Self;
-
-			fn bitand(self, other: Kind) -> Self {
-				match other {
-					$(
-						Kind::$id => Self(self.0 & $mask)
-					),*
-				}
-			}
-		}
-
-		impl std::ops::BitAndAssign<Kind> for KindSet {
-			fn bitand_assign(&mut self, other: Kind) {
-				match other {
-					$(
-						Kind::$id => self.0 &= $mask
-					),*
-				}
-			}
-		}
-
-		impl std::ops::BitAnd for KindSet {
-			type Output = Self;
-
-			fn bitand(self, other: Self) -> Self {
-				Self(self.0 & other.0)
-			}
-		}
-
-		impl std::ops::BitAndAssign for KindSet {
-			fn bitand_assign(&mut self, other: Self) {
-				self.0 &= other.0
-			}
-		}
-
-		impl From<Kind> for KindSet {
-			fn from(value: Kind) -> Self {
-				match value {
-					$(
-						Kind::$id => Self($mask)
-					),*
-				}
-			}
-		}
-
-		impl<'a> IntoIterator for &'a KindSet {
-			type IntoIter = KindSetIter;
-			type Item = Kind;
-
-			fn into_iter(self) -> KindSetIter {
-				self.iter()
-			}
-		}
-
-		impl IntoIterator for KindSet {
-			type IntoIter = KindSetIter;
-			type Item = Kind;
-
-			fn into_iter(self) -> KindSetIter {
-				self.iter()
-			}
-		}
-
-		impl fmt::Display for KindSet {
-			fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-				for (i, kind) in self.into_iter().enumerate() {
-					if i > 0 {
-						f.write_str(", ")?;
-					}
-
-					kind.fmt(f)?;
-				}
-
-				Ok(())
-			}
-		}
-
-		#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-		pub struct KindSetIter(u8);
-
-		impl Iterator for KindSetIter {
-			type Item = Kind;
-
-			fn size_hint(&self) -> (usize, Option<usize>) {
-				let len = self.0.count_ones() as usize;
-				(len, Some(len))
-			}
-
-			fn next(&mut self) -> Option<Kind> {
-				$(
-					if self.0 & $mask != 0 {
-						self.0 &= !$mask;
-						return Some(Kind::$id)
-					}
-				)*
-
-				None
-			}
-		}
-
-		impl std::iter::FusedIterator for KindSetIter {}
-		impl std::iter::ExactSizeIterator for KindSetIter {}
-	};
-}
-
-kind_set! {
-	Null (NULL):       0b000001,
-	Boolean (BOOLEAN): 0b000010,
-	Number (NUMBER):   0b000100,
-	String (STRING):   0b001000,
-	Array (ARRAY):     0b010000,
-	Object (OBJECT):   0b100000
-}
 
 /// JSON Value.
 ///
